requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9

opencv-python==4.10.0.84
torch==2.3.1
numpy==1.26.4

transformers==4.46.3
